refactor: split data serializable to other files
